update : support write & read. pass testing of write or read alone

+++ src/main/java/org/diskqueue/storage/FileCleanupDeleter.java
@@ -10,21 +10,23 @@ import java.util.concurrent.locks.ReentrantLock;
 
 /**
  * Disk file clean up thread.
- *
+ * <p>
  * Delete files on Linux file system with ext3 or later. ext3 will approximately block
  * for a long time while removing a big normal file. discuss at :
- *
+ * <p>
  * http://serverfault.com/questions/128012/how-to-make-rm-faster-on-ext3-linux
- *
+ * <p>
  * therefore, we provide a seperate standalone thread to acompulish
- *
  */
 public class FileCleanupDeleter extends Hypervisor {
+    private static final boolean onlyMarkedRemove = true;
+
     // tasks of file is being deleted
     private BlockingQueue<File> toBeDeleted = new LinkedBlockingQueue<>();
     // task complete notifier
     private final ReentrantLock lock = new ReentrantLock();
     private final Condition wait = lock.newCondition();
+
 
     FileCleanupDeleter() {
         super("Clean-Thread-Deleter", true);
@@ -34,10 +36,11 @@ public class FileCleanupDeleter extends Hypervisor {
     protected boolean execute() {
         File delete = null;
         try {
-            for (;;) {
+            for (; ; ) {
                 delete = toBeDeleted.take();
                 lock.lock();
-                delete.delete();
+                if (!onlyMarkedRemove)
+                    delete.delete();
                 wait.signalAll();
                 lock.unlock();
             }
@@ -49,8 +52,13 @@ public class FileCleanupDeleter extends Hypervisor {
     }
 
     public boolean asyncDelete(File[] files, boolean waitCompleted) {
-        for (File file : files)
-            toBeDeleted.offer(file);
+        // rename first
+        for (File file : files) {
+            File renamed = new File(String.format("%s.deleted", file.getAbsolutePath()));
+            if (!file.renameTo(renamed))
+                System.err.println("data file rename failed : " + renamed.getAbsoluteFile());
+            toBeDeleted.offer(renamed);
+        }
 
         // wait all above inserted file elements for complete
         if (waitCompleted) {
